requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.0
fonttools==4.50.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-resources==6.4.0
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
openai==1.21.2
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pydantic==2.7.0
pydantic-core==2.18.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
tqdm==4.66.2
typing-extensions==4.11.0
zipp==3.18.1
